PubMed frequency list for search query development
This short python script allows you to take a long list of potential search terms and test them as searches in PubMed. The results allow you to determine if any potential search term is too noisy or has zero hits. This is an intermediate step to generating an efficient search query for a systematic review.

You will need to have a PubMed API, which is as easy as cut and paste:
https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/

Hat tip to Don Sizemore at UNC Odum Institute for XML unpacking assist.