List of drugs name sorted by brand vs. generic
This is list of 600+ drug names for opioid analgesics, barbiturates, benzodiazepines, and nootropic stimulants. The flag column identifies whether the name string is likely to be a brand name or a generic name. This may be a useful tool for data cleaning of EHR, EMR, claims data, PDMP data, electronic health records, pharmacy data, etc.
The list has not been manually curated, so use at your risk.